+++ src/demo/entity/main.ts
@@ -1,0 +1,9 @@
+import { Rester } from '../../main';
+import { AccessEntity } from './access.entity';
+import './access.view';
+
+const rester = new Rester();
+
+rester.addEntities({ entities: [AccessEntity] });
+
+rester.bootstrap();